Reorganized kncloudevents retries code
